smart_assistant/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
TTS
torch
torchaudio
soundfile
numpy
